move etcd tests together, rename uptime status to felix_uptime, minor changes
