+++ calico/openstack/test/test_plugin_etcd.py
@@ -795,18 +795,18 @@ class TestPluginEtcd(lib.Lib, unittest.TestCase):
         """
         # Add arbitrary status keys to accumulated etcd db
         self.etcd_data[STATUS_DIR + "/vm_47/status"] = {"status_time": "2015-08-14T10:37:54"}
-        self.etcd_data[STATUS_DIR + "/vm_47/uptime"] = 12
-        self.etcd_data[STATUS_DIR + "/machine_123/uptime"] = 17
+        self.etcd_data[STATUS_DIR + "/vm_47/felix_uptime"] = 12
+        self.etcd_data[STATUS_DIR + "/machine_123/felix_uptime"] = 17
         self.etcd_data[STATUS_DIR + "/VMachine/status"] = {"status_time": "1994-06-09T05:13:18"}
 
         expected_calls = [mock.call(None, TestIfAgentState(host="vm_47",
                                                            start_flag=True,
                                                            agent_type=mech_calico.AGENT_TYPE_FELIX,
-                                                           binary='')),
+                                                           binary='felix-calico-agent')),
                           mock.call(None, TestIfAgentState(host="machine_123",
                                                            start_flag=True,
                                                            agent_type=mech_calico.AGENT_TYPE_FELIX,
-                                                           binary=''))]
+                                                           binary='felix-calico-agent'))]
 
         self.driver.db = mock.Mock()
         self.driver._client = self.client
@@ -827,7 +827,7 @@ class TestPluginEtcd(lib.Lib, unittest.TestCase):
         create_or_update_agent = self.driver.db.create_or_update_agent
 
         # Test status creation
-        status_create = FakeResponse(key=STATUS_DIR+"/VM_1/uptime",
+        status_create = FakeResponse(key=STATUS_DIR+"/VM_1/felix_uptime",
                                      newKey=True)
         self.driver._handle_status_update(status_create)
         create_agent_state = TestIfAgentState(host="VM_1", start_flag=True)
@@ -836,7 +836,7 @@ class TestPluginEtcd(lib.Lib, unittest.TestCase):
         create_or_update_agent.reset_mock()
 
         # Test status update
-        status_update = FakeResponse(key=STATUS_DIR+"/VM_2/uptime")
+        status_update = FakeResponse(key=STATUS_DIR+"/VM_2/felix_uptime")
         self.driver._handle_status_update(status_update)
         update_agent_state = TestIfAgentState(host="VM_2", start_flag=False)
         create_or_update_agent.assert_called_once(self.driver._db_context,
@@ -844,13 +844,13 @@ class TestPluginEtcd(lib.Lib, unittest.TestCase):
         create_or_update_agent.reset_mock()
 
         # Test deleted status
-        status_delete = FakeResponse(key=STATUS_DIR+"/VM_3/uptime",
+        status_delete = FakeResponse(key=STATUS_DIR+"/VM_3/felix_uptime",
                                      action="delete")
         self.driver._handle_status_update(status_delete)
         self.assertFalse(create_or_update_agent.called)
 
         # Test expired status
-        status_expire = FakeResponse(key=STATUS_DIR+"/VM_4/uptime",
+        status_expire = FakeResponse(key=STATUS_DIR+"/VM_4/felix_uptime",
                                      action="expire")
         self.driver._handle_status_update(status_expire)
         self.assertFalse(create_or_update_agent.called)
